+++ source/Nevermore.IntegrationTests/Model/StringTinyTypeIdTypeHandler.cs
@@ -1,8 +1,8 @@
-﻿using System;
+﻿#nullable enable
+using System;
 using System.Data;
 using System.Data.Common;
 using Nevermore.Advanced.TypeHandlers;
-using Nevermore.Mapping;
 
 namespace Nevermore.IntegrationTests.Model
 {
@@ -32,29 +32,6 @@ namespace Nevermore.IntegrationTests.Model
             {
                 parameter.Value = null;
             }
-        }
-    }
-
-    class StringTinyTypeIdKeyHandler<T> : IStringBasedPrimitivePrimaryKeyHandler
-        where T : StringTinyType
-    {
-        public Type Type => typeof(T);
-
-        public object? GetPrimitiveValue(object? id)
-        {
-            if (!(id is StringTinyType stringTinyType))
-                throw new ArgumentException($"Expected the id to be a {typeof(T).Name}");
-            return stringTinyType.Value;
-        }
-
-        public object FormatKey(string tableName, int key)
-        {
-            return TinyType<string>.Create<T>($"{tableName}s-{key}");
-        }
-
-        public void SetIdPrefix(Func<(string tableName, int key), string> idPrefix)
-        {
-            throw new NotImplementedException();
         }
     }
 }
