+++ source/Nevermore/Mapping/IIdentityPrimaryKeyHandler.cs
@@ -1,0 +1,5 @@
+namespace Nevermore.Mapping
+{
+    public interface IIdentityPrimaryKeyHandler : IPrimaryKeyHandler
+    {}
+}
