+++ source/Nevermore/Mapping/IPrimitivePrimaryKeyHandler.cs
@@ -1,0 +1,13 @@
+#nullable enable
+using System.Diagnostics.CodeAnalysis;
+
+namespace Nevermore.Mapping
+{
+    public interface IPrimitivePrimaryKeyHandler : IPrimaryKeyHandler
+    {
+        [return: NotNullIfNotNull("id")]
+        object? GetPrimitiveValue(object? id);
+
+        object FormatKey(string tableName, int key);
+    }
+}
