+++ source/Nevermore/Mapping/IStringBasedPrimitivePrimaryKeyHandler.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Nevermore.Mapping
+{
+    public interface IStringBasedPrimitivePrimaryKeyHandler : IPrimitivePrimaryKeyHandler
+    {
+        void SetIdPrefix(Func<(string tableName, int key), string> idPrefix);
+    }
+}
