+++ source/Nevermore/Mapping/IdentityPrimaryKeyHandler.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Nevermore.Mapping
+{
+    public class IdentityPrimaryKeyHandler<T> : IIdentityPrimaryKeyHandler
+    {
+        public Type Type => typeof(T);
+    }
+}
