+++ source/Nevermore/Mapping/IntPrimaryKeyHandler.cs
@@ -1,0 +1,10 @@
+namespace Nevermore.Mapping
+{
+    class IntPrimaryKeyHandler : PrimitivePrimaryKeyHandler<int>
+    {
+        public override object FormatKey(string tableName, int key)
+        {
+            return key;
+        }
+    }
+}
