+++ source/Nevermore/Mapping/LongPrimaryKeyHandler.cs
@@ -1,0 +1,10 @@
+namespace Nevermore.Mapping
+{
+    class LongPrimaryKeyHandler : PrimitivePrimaryKeyHandler<long>
+    {
+        public override object FormatKey(string tableName, int key)
+        {
+            return key;
+        }
+    }
+}
