+++ source/Nevermore/Mapping/PrimitivePrimaryKeyHandler.cs
@@ -1,0 +1,19 @@
+﻿#nullable enable
+using System;
+using System.Diagnostics.CodeAnalysis;
+
+namespace Nevermore.Mapping
+{
+    public abstract class PrimitivePrimaryKeyHandler<T> : IPrimitivePrimaryKeyHandler
+    {
+        public Type Type => typeof(T);
+
+        [return: NotNullIfNotNull("id")]
+        public virtual object? GetPrimitiveValue(object? id)
+        {
+            return id;
+        }
+
+        public abstract object FormatKey(string tableName, int key);
+    }
+}
